Venue Master APP
Description: As more people are opting for staycations over out of town vacations, a new importance has been placed on finding events and activities in your local area.  With Venue Master, inputting your zip code and desired radius of travel returns your local venues and events that you can save.

Motivation: As individuals, we are all making an effort to be more social and get the full experience our towns have to offer.  Our goal was to create a web application that would ask for a simple input and return all the local venues and the corresponding events in one seamless interface.

Result: Using the Ticket Master API combined with the Google Maps API, we were not only able to query a database of events in a predetermined location/radius, but find their hosting venues and pinpoint them on a map.

Improvements: In the future, we would like to create user logins for individuals to create personalized save events list as opposed to the current global saved event list table.  From there, we would like to integrate these user lists with social media so that that event lists can be shared on Facebook and other platforms.